move back into exceptions.py
